First Commit
Basically, it was redeveloped in a modular structure,

Created the basis of auth with jwt token and issued authorization-based jwt token and made transactions with that token.

We had two roles named user and admin and we stored the data of which user these roles belong to with 2 tables (user, user-role) on the database side. After the user posts the login method with username and password, the jwt token is generated for the role and can perform transactions with that token.

- Previously, the user_role table stored information about user authorizations.
- Now, authorizations are stored in a dedicated roles table with a modular design.
- Users are marked with role IDs using a foreign key in the users table.
- During authorization checks, we retrieve the user's role ID, map it to the corresponding role in the roles table, and handle controller operations accordingly.

is a continuation of the fake-database-api-with-nestjs project.
Since we no longer have anything to do with fake database, I am continuing it as a different project.